Added local-configs/ dir for configuration sharing.
The local-configs/ dir is intended to be used as a configuration
files sharing, so anybody could check someone else's configuration.
Sharing config files is optional.

The local-configs/config.ini-vufind.ics.muni.cz file is intended to
be the final file for production environment.

All of the others are for the reference only.