Add small displays (mobile) for items view.
Items view for records (at record's homepage) doesn't look good on
small displays (mobile phones), becaouse the table is to wide. This
commit fixes this by adding separate formatting for small displays
(xs and sm in terms of Bootstrap 3).

This could be optimized since there are items listed two times and
hidden for certain displays. So the Ajax calls are called two times,
however, the Apache cache will probably do the good job and this won't
need to be optimized from the speed point of view.