Items view table switch to auto table layout.
The default layout (table-layout) of tables in Boostrap is set to fixed,
so all columns are fixed width. It causes troubles if there are longer
words in table's cell - such words (for example "Prirodovedecka" are
then wrapped incorrectly (last letter on the new line etc.). Better is
to set the table to auto, then the width of columns are set according to
the size of a content and wrapping works better.